rename : 경로 변경
- model => domains
- view => InputView
