+++ src/domains/Lotto.js
